updates to auth, refactor, clean-up
