--- travel-event-host-app/src/app/users/[id]/styles.module.css
@@ -1,27 +1,0 @@
-
-.userAvatar {
-  position: relative !important;
-  width: 80px !important;
-  height: 80px !important;
-}
-
-@media only screen and (min-width: 600px) {
-  .userAvatar {
-    width: 100px !important;
-    height: 100px !important;
-  }
-}
-
-@media only screen and (min-width: 960px) {
-  .userAvatar {
-    width: 140px !important;
-    height: 140px !important;
-  }
-}
-
-@media only screen and (min-width: 1440px) {
-  .userAvatar {
-    width: 240px !important;
-    height: 240px !important;
-  }
-}
