+++ travel-event-host-app/src/components/auth-dialog/sign-up-fields/SignUpFields.tsx
@@ -1,8 +1,8 @@
 import theme from '@/app/theme';
 import { Box, Typography } from '@mui/material';
 import { ChangeEventHandler } from 'react';
+import { CustomTextField, StyledFormFieldSection } from '../../custom-fields/CustomFields';
 import { ErrorComponent } from '../ErrorComponent/ErrorComponent';
-import { CustomTextField, StyledFormFieldSection } from '../custom-fields/CustomFields';
 
 export const SignUpFields = (
   handleInputChanged: ChangeEventHandler<HTMLInputElement | HTMLTextAreaElement> | undefined,
